requirements.txt
Django==3.2.6
djangorestframework==3.12.4
psycopg2-binary==2.8.5
Pillow==8.3.1
python-decouple==3.4
djangorestframework_simplejwt==4.4.0